# Performance: Cache `missingSupport` and refactor `Detector`