--- a/lib/BrowserSelection.js
+++ b/lib/BrowserSelection.js
@@ -166,16 +166,29 @@ export default class BrowserSelection {
     return result;
   }
 
+  /** @readonly */
+  static #missingSupportCache = new Map();
+
   /**
    * @see BrowserSelection.compileBrowserSupport
    * @param {ConstructorParameters<typeof this>} constructorParameters
    * @return {MissingSupportResult} `{browsers, features}`
    */
   static missingSupport(...constructorParameters) {
-    const selection = new BrowserSelection(...constructorParameters);
-    return {
+    const [query, path] = constructorParameters;
+    // browserslist only uses `path` if `query` is not provided.
+    const key = query ? JSON.stringify(query) : path;
+
+    if (this.#missingSupportCache.has(key)) {
+      return this.#missingSupportCache.get(key);
+    }
+
+    const selection = new BrowserSelection(query, path);
+    const result = {
       browsers: selection.list(),
       features: selection.compileBrowserSupport(),
     };
+    this.#missingSupportCache.set(key, result);
+    return result;
   }
 }