--- a/lib/BrowserSelection.js
+++ b/lib/BrowserSelection.js
@@ -166,7 +166,7 @@ export default class BrowserSelection {
     return result;
   }
 
-  /** @readonly */
+  /** @type {Map<string, MissingSupportResult>} @readonly */
   static #missingSupportCache = new Map();
 
   /**
@@ -179,7 +179,7 @@ export default class BrowserSelection {
     // browserslist only uses `path` if `query` is not provided.
     const key = query ? JSON.stringify(query) : path;
 
-    if (this.#missingSupportCache.has(key)) {
+    if (key && this.#missingSupportCache.has(key)) {
       return this.#missingSupportCache.get(key);
     }
 
@@ -188,7 +188,7 @@ export default class BrowserSelection {
       browsers: selection.list(),
       features: selection.compileBrowserSupport(),
     };
-    this.#missingSupportCache.set(key, result);
+    if (key) this.#missingSupportCache.set(key, result);
     return result;
   }
 }