--- a/lib/Detector.js
+++ b/lib/Detector.js
@@ -73,13 +73,14 @@ export default class Detector {
    * @param {(FeatureKeys & string)[]} featureList an array of feature slugs (see caniuse-db)
    */
   constructor(featureList) {
-    /** @type {[FeatureKeys, RuleCheck][]} */
-    this.features = featureList
-      .filter((featureName) => FEATURES[featureName] != null)
-      .map((featureName) => {
-        const feature = FEATURES[featureName];
-        return [featureName, normaliseFeature(feature)];
-      });
+    /** @type {Map<FeatureKeys, RuleCheck>} */
+    this.features = new Map();
+    for (const featureName of featureList) {
+      const feature = FEATURES[featureName];
+      if (feature != null) {
+        this.features.set(featureName, normaliseFeature(feature));
+      }
+    }
     /** @type {(FeatureKeys & string)[]} */
     this.ignore = [];
   }
@@ -98,7 +99,7 @@ export default class Detector {
     switch (option) {
       case DISABLE_FEATURE_COMMENT: {
         if (value === '') {
-          this.ignore = this.features.map(([featureName]) => featureName);
+          this.ignore = [...this.features.keys()];
         } else {
           for (const feat of value.split(',')) {
             /** @type {any} */
@@ -135,9 +136,10 @@ export default class Detector {
         return;
       }
 
-      const detectedFeatures = this.features.filter(([, ruleCheck]) => ruleCheck(child));
-      for (const [featureName] of detectedFeatures) {
-        callback({ usage: child, feature: featureName, ignore: this.ignore });
+      for (const [feature, ruleCheck] of this.features) {
+        if (ruleCheck(child)) {
+          callback({ usage: child, feature, ignore: this.ignore });
+        }
       }
 
       if (child.type !== 'decl') {